IPreparedStatment: add setBlob()
The Android layer has been reworked with two different implementations
for IPreparedStatement dependending on the query type. Setting blobs
as arguments is then only supported with update, insert and delete but
not with select (because the underlying SQLiteDatabase#rawQuery() does
not support other arguments than String and blob cannot be converted to
that.